Move mock request_airdrop_transaction into drone/, closer to the real impl
