+++ drone/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod drone;
+pub mod drone_mock;
