+++ wallet/src/wallet.rs
@@ -8,7 +8,7 @@ use serde_json::json;
 use solana_budget_api;
 use solana_budget_api::budget_transaction::BudgetTransaction;
 #[cfg(test)]
-use solana_client::rpc_mock::{request_airdrop_transaction, MockRpcClient as RpcClient};
+use solana_client::rpc_mock::MockRpcClient as RpcClient;
 #[cfg(not(test))]
 use solana_client::rpc_request::RpcClient;
 use solana_client::rpc_request::{get_rpc_request_str, RpcRequest};
@@ -16,6 +16,8 @@ use solana_client::rpc_signature_status::RpcSignatureStatus;
 #[cfg(not(test))]
 use solana_drone::drone::request_airdrop_transaction;
 use solana_drone::drone::DRONE_PORT;
+#[cfg(test)]
+use solana_drone::drone_mock::request_airdrop_transaction;
 use solana_sdk::bpf_loader;
 use solana_sdk::hash::Hash;
 use solana_sdk::loader_transaction::LoaderTransaction;
